+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.98.49", date: "2026-04-05" };
+  return { version: "8.98.50", date: "2026-04-05" };
 }